Updates build configuration for Gradle
This commit updates the build configuration files to align with a Gradle upgrade.

It configures duplicate strategy to exclude for Jar, Tar, and Zip tasks.
Removes unnecessary source set configurations and test configurations in build.gradle files.
Renames packages for test java classes in server module to align with the new structure.

Signed-off-by: Nico Piel <nico.piel@hotmail.de>